move demographic chart to own file
