+++ src/components/CohortBuilder/Summary/index.js
@@ -5,11 +5,11 @@ import { withTheme } from 'emotion-theming';
 import { demographicPiesMock, topDiagnosesBarMock, studiesBarMock } from './mock';
 import Card from 'uikit/Card';
 import MultiHeader from 'uikit/Multicard/MultiHeader';
-import { CardWrapper } from 'uikit/Card/styles';
 import { Col, Row } from 'react-grid-system';
-import Pie from 'chartkit/components/Pie';
 import HorizontalBar from 'chartkit/components/HorizontalBar';
-import DataProvider from '../DataProvider';
+import QueriesResolver from '../QueriesResolver';
+import { withApi } from 'services/api';
+import DemographicChart, { demographicQuery } from './DemographicChart';
 
 const mostFrequentDiagnosisTooltip = data => {
   const participants = data.familyMembers + data.probands;
@@ -39,15 +39,6 @@ const CardSlot = styled(Card)`
   height: 305px;
 `;
 
-const CardSlotPies = styled(CardWrapper)`
-  height: 305px;
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 10px 10px;
-`;
-
 const LongCard = styled(Card)`
   height: 100%;
 `;
@@ -55,7 +46,10 @@ const LongCard = styled(Card)`
 const md = 4;
 const lg = 4;
 
-const enhance = compose(withTheme);
+const enhance = compose(
+  withApi,
+  withTheme,
+);
 
 const multiHeader = (
   <MultiHeader
@@ -63,94 +57,75 @@ const multiHeader = (
   />
 );
 
-const Summary = ({ theme, sqon }) => (
-  <Row nogutter>
-    <Col sm={12} md={9} lg={9}>
-      <Row nogutter>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlot title="Overall Survival" />
-        </Col>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlot title={multiHeader}>
-            <HorizontalBar
-              data={studiesBarMock}
-              indexBy="label"
-              keys={['probands', 'familyMembers']}
-              tooltipFormatter={studiesToolTip}
-              sortBy={sortDescParticipant}
-              tickInterval={4}
-              colors={[theme.chartColors.blue, theme.chartColors.purple]}
-              xTickTextLength={28}
-              legends={[
-                { title: 'Probands', color: theme.chartColors.blue },
-                { title: 'Family Members', color: theme.chartColors.purple },
-              ]}
-              padding={0.5}
-            />
-          </CardSlot>
-        </Col>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlot title="Most Frequent Diagnoses">
-            <HorizontalBar
-              style={{ maxWidth: '100px' }}
-              data={topDiagnosesBarMock}
-              indexBy="label"
-              keys={['probands', 'familyMembers']}
-              tooltipFormatter={mostFrequentDiagnosisTooltip}
-              sortByValue={true}
-              tickInterval={4}
-              colors={[theme.chartColors.blue, theme.chartColors.purple]}
-              xTickTextLength={28}
-              legends={[
-                { title: 'Probands', color: theme.chartColors.blue },
-                { title: 'Family Members', color: theme.chartColors.purple },
-              ]}
-              padding={0.5}
-            />
-          </CardSlot>
-        </Col>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlotPies>
-            <Pie
-              style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
-              title={'Gender'}
-              data={demographicPiesMock.gender}
-              colors={[theme.chartColors.orange, '#FFFFFF']}
-            />
-            <Pie
-              style={{ height: '42%', width: '50%', marginBottom: '10px', marginTop: '5px' }}
-              title={'Ethnicity'}
-              data={demographicPiesMock.ethnicity}
-              colors={[theme.chartColors.darkblue, '#FFFFFF']}
-            />
-            <Pie
-              style={{ height: '42%', width: '50%' }}
-              title={'Race'}
-              data={demographicPiesMock.race}
-              colors={[theme.chartColors.lightpurple, '#FFFFFF']}
-            />
-            <Pie
-              style={{ height: '42%', width: '50%' }}
-              title={'Family Composition'}
-              data={demographicPiesMock.familyComposition}
-              colors={[theme.chartColors.lightblue, '#FFFFFF']}
-            />
-          </CardSlotPies>
-        </Col>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlot title="File Breakdown" />
-        </Col>
-        <Col sm={12} md={md} lg={lg}>
-          <CardSlot title="Age at Diagnosis" />
-        </Col>
-      </Row>
-    </Col>
-    <Col sm={12} md={3} lg={3}>
-      <LongCard title="Phenotypes">
-        <pre>{JSON.stringify(sqon, null, 2)}</pre>
-      </LongCard>
-    </Col>
-  </Row>
+const Summary = ({ theme, sqon, api }) => (
+  <QueriesResolver api={api} queries={[demographicQuery({ ...{ sqon } })]}>
+    {({ loading, results }) => {
+      return (
+        <Row nogutter>
+          <Col sm={12} md={9} lg={9}>
+            <Row nogutter>
+              <Col sm={12} md={md} lg={lg}>
+                <CardSlot title="Overall Survival" />
+              </Col>
+              <Col sm={12} md={md} lg={lg}>
+                <CardSlot title={multiHeader}>
+                  <HorizontalBar
+                    data={studiesBarMock}
+                    indexBy="label"
+                    keys={['probands', 'familyMembers']}
+                    tooltipFormatter={studiesToolTip}
+                    sortBy={sortDescParticipant}
+                    tickInterval={4}
+                    colors={[theme.chartColors.blue, theme.chartColors.purple]}
+                    xTickTextLength={28}
+                    legends={[
+                      { title: 'Probands', color: theme.chartColors.blue },
+                      { title: 'Family Members', color: theme.chartColors.purple },
+                    ]}
+                    padding={0.5}
+                  />
+                </CardSlot>
+              </Col>
+              <Col sm={12} md={md} lg={lg}>
+                <CardSlot title="Most Frequent Diagnoses">
+                  <HorizontalBar
+                    style={{ maxWidth: '100px' }}
+                    data={topDiagnosesBarMock}
+                    indexBy="label"
+                    keys={['probands', 'familyMembers']}
+                    tooltipFormatter={mostFrequentDiagnosisTooltip}
+                    sortByValue={true}
+                    tickInterval={4}
+                    colors={[theme.chartColors.blue, theme.chartColors.purple]}
+                    xTickTextLength={28}
+                    legends={[
+                      { title: 'Probands', color: theme.chartColors.blue },
+                      { title: 'Family Members', color: theme.chartColors.purple },
+                    ]}
+                    padding={0.5}
+                  />
+                </CardSlot>
+              </Col>
+              <Col sm={12} md={md} lg={lg}>
+                <DemographicChart data={demographicPiesMock} />
+              </Col>
+              <Col sm={12} md={md} lg={lg}>
+                <CardSlot title="File Breakdown" />
+              </Col>
+              <Col sm={12} md={md} lg={lg}>
+                <CardSlot title="Age at Diagnosis" />
+              </Col>
+            </Row>
+          </Col>
+          <Col sm={12} md={3} lg={3}>
+            <LongCard title="Phenotypes">
+              <pre>{JSON.stringify(sqon, null, 2)}</pre>
+            </LongCard>
+          </Col>
+        </Row>
+      );
+    }}
+  </QueriesResolver>
 );
 
 export default enhance(Summary);
